Move health kit controller test to specific suite
