+++ paso/pasoTests/pasoTests.swift
@@ -15,23 +15,6 @@ class pasoTests: XCTestCase {
         // Put teardown code here. This method is called after the invocation of each test method in the class.
     }
 
-    func testFailureToGetStepsTriggersAlert() async throws {
-        let healthKitController = HealthKitController()
-        do {
-            // Property 'shouldAlert' is false by default.
-            XCTAssert(healthKitController.shouldAlert == false,
-                      "Controller's alert property should be false by default.")
-            // Without user interaction the permissions will not be changed,
-            // and the call to update the step count will fail as a result.
-            try await healthKitController.authorizeHealthKit()
-            try await healthKitController.updatePublishedStepCount()
-            XCTAssert(healthKitController.shouldAlert == true,
-                      "Controller's alert property should be true if step count is not read.")
-        } catch {
-            XCTFail()
-        }
-    }
-
     func testLaunchPerformance() throws {
         // This is an example of a performance test case.
         self.measure {
